Return buffer and delegate buffer's process management to make-comint-in-buffer
By returning `buffer` we solve an issue in Doom Emacs which wasn't able
to properly manage the repl buffer when open through `SPC o r`.

Some bits seemed to be redundant since `make-comint-in-buffer` was
already taking care of checking whethere there is a comint process
running assigned to the current buffer.