Fix buffer not being automatically closed on process exit
Doom Emacs figures out whether to re-use or not the existing repl buffer
by calling `buffer-live-p` on the buffer. If it's alive, no further
action would take place.

However, nowadays when purescript repl exits, the buffer is still alive,
except, the process is not.

To cater for the case where the process exists, e.g. by entering `:q` in
the purescript repl, we would close the buffer.

With that fix, `buffer-live-p` would return `nil` and start a new process.